updated creation of books to line up with requirements

+++ src/main/java/edu/fiu/group24/group_24_project/Authors/AuthorController.java
@@ -21,7 +21,7 @@ public class AuthorController {
     private PublisherRepository publisherRepository;
 
     @GetMapping("/{authorId}/books")
-    public ResponseEntity<List<Book>> getAllBooksByAuthor(@PathVariable Integer authorId){
+    public ResponseEntity<List<Book>> getAllBooksByAuthor(@PathVariable Long authorId){
         if (!authorRepository.existsById(authorId)) {
             return ResponseEntity.notFound().build();
         }
